New codon optimization, better removal of TMP directories, better con…

+++ pangenome/bin/FELIX/pgg_edge_multifasta.pl
@@ -26,8 +26,8 @@ use strict;
 use warnings;
 use Getopt::Std;
 use File::Basename;
-getopts ('j:I:RSALlDhb:B:m:p:P:a:g:t:M:s:T:Vk:FfC:Q:N:Xr:W:G:');
-our ($opt_j,$opt_I,$opt_S,$opt_A,$opt_L,$opt_l,$opt_D,$opt_h,$opt_b,$opt_m,$opt_p,$opt_P,$opt_a,$opt_g,$opt_t,$opt_M,$opt_r,$opt_R,$opt_B,$opt_s,$opt_T,$opt_V,$opt_k,$opt_F,$opt_f,$opt_C,$opt_Q,$opt_N,$opt_X,$opt_W,$opt_G);
+getopts ('j:I:ORSALlDhb:B:m:p:P:a:g:t:M:s:T:Vk:FfC:Q:N:Xr:W:G:');
+our ($opt_j,$opt_I,$opt_O,$opt_S,$opt_A,$opt_L,$opt_l,$opt_D,$opt_h,$opt_b,$opt_m,$opt_p,$opt_P,$opt_a,$opt_g,$opt_t,$opt_M,$opt_r,$opt_R,$opt_B,$opt_s,$opt_T,$opt_V,$opt_k,$opt_F,$opt_f,$opt_C,$opt_Q,$opt_N,$opt_X,$opt_W,$opt_G);
 
 ## use boolean logic:  TRUE = 1, FALSE = 0
 
@@ -65,12 +65,14 @@ my $medoids_path;
 my $single_cores;
 my $topology_file;
 my $strip_version = 0;
+my $codon_opt = 0;
+my $codon_opt_file;
 my $qsub_queue = "himem";
 my $wall_time_limit = "24:00:00"; #set qsub wall time limit to 24 hours by default
 my $mem_req = "2gb"; #set qsub memory minimum requirement by default to 2 Gbyte
 my $MAX_ALIGN_EDGE = 30000; #maximum length of an edge that we are willing to multiple sequence align to
 my $MIN_ALIGN_EDGE = 5; #minimum length of an edge that we are willing to multiple sequence align to
-my $MIN_PGG_PID = 98; #minimum threshold to perform multiple sequence alignment of genomes - must be below this threshold to align
+my $MIN_PGG_PID = 101; #minimum threshold to perform multiple sequence alignment of genomes - must be below this threshold to align (set to 98 if fewer alignments desired)
 
 if ($opt_W) { # should really check time format
     $wall_time_limit = $opt_W;
@@ -93,6 +95,7 @@ if ($opt_s) {$single_cores = $opt_s;} else {$single_cores = "";}
 if ($opt_X) {$no_stats = 1;} else {$no_stats = 0;}
 if ($opt_F) {$use_multifasta = 1;} else {$use_multifasta = 0;}
 if ($opt_f) {$write_multifasta = 1;} else {$write_multifasta = 0;}
+if ($opt_O) {$codon_opt = 1;} else {$codon_opt = 0;}
 if ($opt_R) {$remake_files = 1;} else {$remake_files = 0;}
 if ($opt_V) {$strip_version = 1;} else {$strip_version = 0;}
 if ($opt_A) {$compute_all = 1;} else {$compute_all = 0;} # flag to compute statistics for all
@@ -174,6 +177,12 @@ if ($opt_r) {
     }
 } else {
     $rscript_path = "/usr/local/bin/Rscript";
+}
+if ($codon_opt) {
+    if ($target_id eq "") {
+	die "Cannot specify codon optimization output (-O) without providing a target genome (-t)\n";
+    }
+    $codon_opt_file = "$basedir/$target_id" . "_codon_opt.txt";
 }
 
 my $num_size_one_clus = 0;
@@ -1407,21 +1416,30 @@ sub process_matchtable {
 					    }
 					}
 					close(STATSFILE);
-					if ($keep_divergent_alignments) {
+					print STDERR "$target_id\t$cluster_id\t$median_target\t$max_all\t$cols_target\t$max_cols_all\n" if ($DEBUG);
+					my $diverged_alignment_flag = 0;
+					if ($median_target > $max_all) {
+					    print DIFFFILE "$target_id\t$feat_hash{$feat_name}->{'contig'}\tdivergent_clus_allele\t$feat_hash{$feat_name}->{'5p'}\t$feat_hash{$feat_name}->{'3p'}\t$feat_hash{$feat_name}->{'len'}\t$feat_name\n";
+					    $distant_clus_alle{$target_id}++;
+					    $diverged_alignment_flag = 1;
+					}
+					if ($cols_target > $max_cols_all) {
+					    print DIFFFILE "$target_id\t$feat_hash{$feat_name}->{'contig'}\tdivergent_columns_clus_allele\t$feat_hash{$feat_name}->{'5p'}\t$feat_hash{$feat_name}->{'3p'}\t$feat_hash{$feat_name}->{'len'}\t$feat_name\n";
+					    $column_clus_alle{$target_id}++;
+					    $diverged_alignment_flag = 1;
+					}
+					if (!$diverged_alignment_flag) {
+					    print DIFFFILE "$target_id\t$feat_hash{$feat_name}->{'contig'}\tconserved_clus_allele\t$feat_hash{$feat_name}->{'5p'}\t$feat_hash{$feat_name}->{'3p'}\t$feat_hash{$feat_name}->{'len'}\t$feat_name\n";
+					}
+					if ($codon_opt && $diverged_alignment_flag) {
+					    `sed -e "s/$target_id/$cluster_id/" $seq_file >> $codon_opt_file`;
+					}
+					if ($keep_divergent_alignments && $diverged_alignment_flag) {
 					    `mv $combined_file $keep_divergent_alignments/cluster_$target_id.$cluster_id.afa`;
 					    `mv $com_stats_file $keep_divergent_alignments/cluster_$target_id.$cluster_id.stats`;
 					    `rm $seq_file`;
 					} else {
 					    `rm $seq_file $combined_file $com_stats_file`;
-					}
-					print STDERR "$target_id\t$cluster_id\t$median_target\t$max_all\t$cols_target\t$max_cols_all\n" if ($DEBUG);
-					if ($median_target > $max_all) {
-					    print DIFFFILE "$target_id\t$feat_hash{$feat_name}->{'contig'}\tdivergent_clus_allele\t$feat_hash{$feat_name}->{'5p'}\t$feat_hash{$feat_name}->{'3p'}\t$feat_hash{$feat_name}->{'len'}\t$feat_name\n";
-					    $distant_clus_alle{$target_id}++;
-					}
-					if ($cols_target > $max_cols_all) {
-					    print DIFFFILE "$target_id\t$feat_hash{$feat_name}->{'contig'}\tdivergent_columns_clus_allele\t$feat_hash{$feat_name}->{'5p'}\t$feat_hash{$feat_name}->{'3p'}\t$feat_hash{$feat_name}->{'len'}\t$feat_name\n";
-					    $column_clus_alle{$target_id}++;
 					}
 				    }
 				}
@@ -2258,20 +2276,27 @@ sub process_pgg {
 						}
 					    }
 					    close(STATSFILE);
-					    if ($keep_divergent_alignments) {
-						`mv $combined_file $keep_divergent_alignments/edge_$target_id.$edge_id.afa`;
-						`rm $seq_file $com_stats_file`;
-					    } else {
-						`rm $seq_file $combined_file $com_stats_file`;
-					    }
 					    print STDERR "$target_id\t$edge_id\t$median_target\t$max_all\t$cols_target\t$max_cols_all\n" if ($DEBUG);
+					    my $diverged_alignment_flag = 0;
 					    if ($median_target > $max_all) {
 						print DIFFFILE "$target_id\t$edge_hash{$feat_name}->{'contig'}\tdivergent_edge_allele\t$edge_hash{$feat_name}->{'5p'}\t$edge_hash{$feat_name}->{'3p'}\t$edge_hash{$feat_name}->{'len'}\t$edge_name\n";
 						$distant_edge_alle{$target_id}++;
+						$diverged_alignment_flag = 1;
 					    }
 					    if ($cols_target > $max_cols_all) {
 						print DIFFFILE "$target_id\t$edge_hash{$feat_name}->{'contig'}\tdivergent_columns_edge_allele\t$edge_hash{$feat_name}->{'5p'}\t$edge_hash{$feat_name}->{'3p'}\t$edge_hash{$feat_name}->{'len'}\t$edge_name\n";
 						$column_edge_alle{$target_id}++;
+						$diverged_alignment_flag = 1;
+					    }
+					    if (!$diverged_alignment_flag) {
+						print DIFFFILE "$target_id\t$edge_hash{$feat_name}->{'contig'}\tconserved_edge_allele\t$edge_hash{$feat_name}->{'5p'}\t$edge_hash{$feat_name}->{'3p'}\t$edge_hash{$feat_name}->{'len'}\t$edge_name\n";
+					    }
+					    if ($keep_divergent_alignments) {
+						`mv $combined_file $keep_divergent_alignments/edge_$target_id.$edge_id.afa`;
+						`mv $com_stats_file $keep_divergent_alignments/edge_$target_id.$edge_id.stats`;
+						`rm $seq_file`;
+					    } else {
+						`rm $seq_file $combined_file $com_stats_file`;
 					    }
 					}
 				    }
@@ -2794,6 +2819,7 @@ Version: $version
      -f: generate multifasta files with all alleles for clusters and edges
      -F: use multifasta files with all alleles for clusters and edges instead of extracting them directly from genome fasta files
      -C: path to the Muslce executable for multiple sequence alignments - default /usr/local/bin/muscle
+     -O: flag indicating a multifasta file of node/cluster/gene sequences which are significantly diverged should be output for codon optimization
      -r: path to the Rscript executable for Rscript scripts - default /usr/local/bin/Rscript
      -Q: name of qsub queue to use for Muscle jobs - default himem
      -G: minimum memory requirement for grid jobs in format (integer)gb where 2gb is the default
